now it works! sinf -> sin

+++ friedmann.js
@@ -577,7 +577,7 @@ function rightMouseClickHandler(event) {
 /* 
 function bananaProjectionMatrixFrontHalf(fovY, aspect, z0)
 {
-	const height = 1 / Math.tanf(fovY / 2);
+	const height = 1 / Math.tan(fovY / 2);
 	const width = height / aspect;
 	return Matrix(width, 0., 0., 0.,
 		0., height, 0., 0.,
@@ -593,8 +593,8 @@ function sphericalRotationZW(d) // знаки поменял в сравнени
   //from values: 
 	return mat4.fromValues(	1, 0, 0, 0,
 					0, 1, 0, 0,
-					0, 0, Math.cosf(d), Math.sinf(d),
-					0, 0, -Math.sinf(d), Math.cosf(d));
+					0, 0, Math.cos(d), Math.sin(d),
+					0, 0, -Math.sin(d), Math.cos(d));
 }
 
 //todo: остальные также
